2018-12-23 db_bench 每分钟统计数据

+++ tools/db_bench_tool.cc
@@ -4258,7 +4258,7 @@ private:
             // test
 
             //if ((num_written) % FLAGS_num_stat == 0) {
-            if ((Env::Default()->NowMicros() - finish_last_) % 60000000 = 0) {
+            if (static_cast<int>(Env::Default()->NowMicros() - finish_last_) % 60000000 = 0) {
                 double now = Env::Default()->NowMicros();
                 double time = now - finish_last_;
                 int64_t ebytes = bytes - bytes_last_;
